+++ tests/unit/test_bolfire_unit.py
@@ -45,19 +45,15 @@ def parameter_values():
 @pytest.fixture
 def bolfire_method(true_param, seed):
     m = simple_gaussian_model(true_param, seed)
-    return elfi.BOLFIRE(m)
+    return elfi.BOLFIRE(m, 10)
 
 
 def test_generate_marginal(bolfire_method):
     assert bolfire_method._generate_marginal().shape == (10, 10)
 
 
-def test_generate_likelihood(bolfire_method, parameter_values):
-    assert bolfire_method._generate_likelihood(parameter_values).shape == (10, 10)
-
-
 def test_generate_training_data(bolfire_method, parameter_values):
-    likelihood = bolfire_method._generate_likelihood(parameter_values)
+    likelihood = np.random.rand(10, 10)
     X, y = bolfire_method._generate_training_data(likelihood, bolfire_method.marginal)
     assert X.shape == (20, 10)
     assert y.shape == (20,)